feat: simplify file structure
