--- x/commitment/client/wasm/types/types.go
@@ -1,1 +1,0 @@
-package types
