--- x/margin/client/wasm/querier/querier.go
@@ -1,26 +1,0 @@
-package querier
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	wasmbindingstypes "github.com/elys-network/elys/wasmbindings/types"
-	"github.com/elys-network/elys/x/margin/keeper"
-)
-
-// Querier handles queries for the Margin module.
-type Querier struct {
-	keeper *keeper.Keeper
-}
-
-func NewQuerier(keeper *keeper.Keeper) *Querier {
-	return &Querier{
-		keeper: keeper,
-	}
-}
-
-func (oq *Querier) HandleQuery(ctx sdk.Context, query wasmbindingstypes.ElysQuery) ([]byte, error) {
-	switch {
-	default:
-		// This handler cannot handle the query
-		return nil, wasmbindingstypes.ErrCannotHandleQuery
-	}
-}
